+++ app/(site)/layout.tsx
@@ -1,4 +1,4 @@
-import "../futbolhoy.css";
+import "../site-shell.css";
 import "../v5-features.css";
 
 export default function SiteLayout({